MMT-1281 move soap request classes to a new module
